# api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bina Academy API",
    "description": "## نظرة عامة | Overview\n\nمجموعة الـ APIs الخاصة بمنصة **Bina Academy** لطلبة التوجيهي (الفرع العلمي والأدبي - الصف 11 و 12).\n\nAPI collection for **Bina Academy** tutoring platform (Tawjihi students - grades 11/12, scientific/literary branches).\n\n---\n\n### خصائص عامة | General Properties\n\n- جميع الردود JSON (`Content-Type: application/json`)\n- جميع أرقام الهواتف بالصيغة الفلسطينية `^05\\d{8}$` (مثال: `0591234567`)\n- التواريخ بصيغة ISO 8601\n- المصادقة بـ JWT Bearer (توكينات الطلاب 30 يوم، توكينات الإدمن 7 أيام)\n- الحد العام للطلبات: 100 طلب/دقيقة لكل IP (حسب `@nestjs/throttler`)\n\n### تدفّق العمل الموصى به | Recommended Mobile Flows\n\n1. **تسجيل طالب جديد**: POST `/auth/register` (الهاتف، كلمة المرور، الاسم الكامل، الصف، الفرع) → تصفح المواد مباشرة.\n2. **دخول طالب مسجل**: POST `/auth/login` (الهاتف وكلمة المرور) → تصفح المواد.\n3. **مشاهدة فيديو**: GET `/subjects/:id/videos` → GET `/videos/:id/stream` → POST `/videos/:id/mark-viewed`\n4. **إدمن**: POST `/auth/admin/login` → استعمل الـ token في كل endpoints الـ admin\n",
    "version": "1.0.0",
    "contact": {
      "name": "Bina Academy Backend Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development"
    },
    {
      "url": "http://10.0.2.2:3000/api",
      "description": "Android emulator (loopback to host)"
    },
    {
...
    }
  ],
...
  "paths": {
...
    "/me": {
      "get": {
        "tags": ["Users"],
        "summary": "بيانات الطالب الحالي والاشتراك الفعّال",
        "description": "يعود بمعلومات الطالب وتلخيص الاشتراك الفعّال (نوع الخطة، تاريخ الانتهاء، عدد الفيديوهات المتاحة لكل مادة). يتطلب توكين STUDENT.\n\nReturns the current student and their active subscription summary. Requires STUDENT JWT.",
        "operationId": "getMe",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "بيانات الطالب",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MeResponse" },
                "example": {
                  "id": "clv123abc",
                  "phone": "0591234567",
                  "name": "أحمد محمود",
                  "grade": "GRADE_12",
                  "branch": "SCIENTIFIC",
                  "parentPhone": "0599876543",
                  "role": "STUDENT",
                  "subscription": {
                    "planType": "MONTHLY",
                    "planName": "اشتراك شهري",
                    "endDate": "2026-07-28T15:00:00.000Z",
                    "isActive": true,
                    "videosPerSubject": 10
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "المستخدم غير موجود (التوكن يشير لمستخدم محذوف)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "statusCode": 404,
                  "message": "المستخدم غير موجود"
                }
              }
            }
          }
        }
      }
    },
    "/me/parent-phone": {
      "put": {
        "tags": ["Users"],
        "summary": "تحديث رقم هاتف ولي الأمر",
        "description": "يضبط `parentPhone` على رقم فلسطيني صالح. يتطلب توكين STUDENT.",
        "operationId": "updateParentPhone",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateParentPhoneDto" },
              "example": { "parentPhone": "0599876543" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "تم التحديث",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "parentPhone": { "type": "string", "example": "0599876543" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/videos/{id}": {
      "get": {
        "tags": ["Videos"],
        "summary": "تفاصيل الدرس وال chapters والفيديوهات المقترحة | Get Lesson Details",
        "description": "يعيد تفاصيل الدرس (الفيديو)، بما في ذلك رابط الـ PDF، معرف التقييم اليومي، المعلم، الفصول/القطع (chapters) مع حالاتها (منجزة، قيد التشغيل، مقفلة)، الفيديوهات ذات الصلة من نفس المادة، والعلامة المائية للطالب.",
        "operationId": "getLesson",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "بيانات الدرس بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "video": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "title": { "type": "string" },
                        "description": { "type": "string", "nullable": true },
                        "streamUrl": { "type": "string", "nullable": true },
                        "pdfUrl": { "type": "string", "nullable": true },
                        "durationSec": { "type": "integer" },
                        "unitName": { "type": "string" },
                        "subHeader": { "type": "string" },
                        "teacherName": { "type": "string", "nullable": true },
                        "dailyQuizId": { "type": "string", "nullable": true }
                      }
                    },
                    "chapters": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "title": { "type": "string" },
                          "startSec": { "type": "integer" },
                          "endSec": { "type": "integer" },
                          "durationText": { "type": "string" },
                          "status": { "type": "string", "enum": ["COMPLETED", "PLAYING", "UNPLAYED", "LOCKED"] }
                        }
                      }
                    },
                    "relatedVideos": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "title": { "type": "string" },
                          "subjectName": { "type": "string" },
                          "teacherName": { "type": "string", "nullable": true },
                          "durationText": { "type": "string" },
                          "unitName": { "type": "string" }
                        }
                      }
                    },
                    "watermark": {
                      "type": "object",
                      "properties": {
                        "name": { "type": "string" },
                        "phone": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "description": "الفيديو غير موجود" }
        }
      }
    },
    "/videos/{id}/stream": {
      "get": {
        "tags": ["Videos"],
        "summary": "الحصول على رابط بث الفيديو + علامة مائية",
        "description": "يرجع `streamUrl` و `watermark` (اسم الطالب ورقم هاتفه) لإظهارها على المشغّل. **مهم**: تحقّق الـ quota (locked) و عدّاد `maxViews` على كل طلب. يتطلب توكين STUDENT.\n\n> ⚠️ **علامة مائية**: اعرض `watermark.name` و `watermark.phone` على المشغّل لمنع إعادة المشاركة.\n\n> ⚠️ **حدّ المشاهدات**: إذا كان المستخدم قد شاهد الفيديو `maxViews` مرات (افتراضياً 3)، سيرد بـ 403 حتى لو الـ quota يسمح.\n\nReturns the stream URL, watermark info, and playback config. Validates quota AND maxViews on every call. Requires STUDENT JWT.",
        "operationId": "getVideoStream",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "example": "vid_001" }
          }
        ],
        "responses": {
          "200": {
            "description": "بيانات البث ومتطلبات المشغّل",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StreamTokenResponse" },
                "example": {
                  "streamUrl": "https://cdn.bina-academy.ps/videos/vid_001/master.m3u8",
                  "watermark": {
                    "name": "أحمد محمود",
                    "phone": "0591234567"
                  },
                  "playbackRates": [0.75, 1, 1.25, 1.5, 2],
                  "qualities": ["auto", "360p", "480p", "720p", "1080p"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": {
            "description": "الاشتراك غير فعّال **أو** الفيديو خارج الحصة **أو** استُنفدت المشاهدات",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "noSub": {
                    "summary": "لا اشتراك فعّال",
                    "value": { "statusCode": 403, "message": "الاشتراك غير فعّال" }
                  },
                  "outOfQuota": {
                    "summary": "الفيديو خارج حصة الخطة",
                    "value": { "statusCode": 403, "message": "هذا الفيديو غير متاح في خطتك" }
                  },
                  "maxViews": {
                    "summary": "تم بلوغ maxViews",
                    "value": { "statusCode": 403, "message": "تم استنفاد عدد المشاهدات المسموح" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "الفيديو غير موجود أو غير منشور",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "statusCode": 404,
                  "message": "الفيديو غير موجود"
                }
              }
            }
          }
        }
      }
    },
    "/videos/{id}/mark-viewed": {
      "post": {
        "tags": ["Videos"],
        "summary": "تأكيد أن الطالب شاهد الفيديو (عدّاد)",
        "description": "يزيد `viewCount` لهذا الطالب على هذا الفيديو بمقدار 1 ويحدّث `lastViewed`. **استدعِه عند اكتمال المشاهدة** أو عند الإغلاق (لاحتساب المشاهدات بدقة). لا يفحص quota أو maxViews هنا — التحقق يحدث في `/stream`. يتطلب توكين STUDENT.\n\nIncrements view count and updates lastViewed timestamp. Call when playback completes (or on app foreground-close) to track accurately.",
        "operationId": "markVideoViewed",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "example": "vid_001" }
          }
        ],
        "responses": {
          "201": {
            "description": "تم تسجيل المشاهدة",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "viewCount": { "type": "integer", "example": 2 },
                    "maxViews": { "type": "integer", "example": 3 },
                    "triggerQuiz": {
                      "type": "object",
                      "nullable": true,
                      "properties": {
                        "title": { "type": "string", "example": "اختبار قصير: الفيزياء" },
                        "subjectId": { "type": "string" },
                        "questions": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": { "type": "string" },
                              "text": { "type": "string" },
                              "options": {
                                "type": "array",
                                "items": { "type": "string" }
                              },
                              "answer": { "type": "string" }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "الفيديو غير موجود",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "statusCode": 404,
                  "message": "الفيديو غير موجود"
                }
              }
            }
          }
        }
      }
    },
    "/videos/{id}/download-token": {
      "get": {
        "tags": ["Videos"],
        "summary": "توكين التنزيل بدون إنترنت",
        "description": "يولّد توكين تنزيل صالحة لمدة `downloadDays` (افتراضياً 7 أيام من DB). **التوكين حالياً placeholder نصي** (`offline_{videoId}_{userId}_{ts}`) — يلزم ربطه بـ Signed URL من CDN لاحقاً.\n\n> ⚠️ **Expiry**: احفظ `expiresAt` محلياً واطلب توكين جديداً قبل انتهائه.\n\nGenerates an offline download token valid for `downloadDays` (default 7). Currently the token is a placeholder string — to be wired to CDN signed URLs.",
        "operationId": "getVideoDownloadToken",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "example": "vid_001" }
          }
        ],
        "responses": {
          "200": {
            "description": "توكين التنزيل وتاريخ الانتهاء",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DownloadTokenResponse" },
                "example": {
                  "token": "offline_vid_001_clv123abc_1719600000000",
                  "expiresAt": "2026-07-05T15:00:00.000Z",
                  "downloadDays": 7
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": {
            "description": "الاشتراك غير فعّال أو الفيديو خارج الحصة",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "statusCode": 403,
                  "message": "هذا الفيديو غير متاح في خطتك"
                }
              }
            }
          },
          "404": {
            "description": "الفيديو غير موجود",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "statusCode": 404,
                  "message": "الفيديو غير موجود"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer.\n\n- توكين STUDENT: يولَّد من `/auth/verify-otp` (صالح 30 يوم، يستخدم `JWT_SECRET`)\n- توكين ADMIN: يولَّد من `/auth/admin/login` (صالح 7 أيام، يستخدم `ADMIN_JWT_SECRET` أو `JWT_SECRET` كاحتياط)\n\n> ⚠️ **مهم**: لا تستخدم توكين STUDENT في endpoints الـ Admin والعكس صحيح — كلاهما سيرد بـ 401."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Validation error - عادةً من class-validator على الحقول المطلوبة/الصيغة",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "statusCode": 400,
              "message": ["phone must match /^05\\d{8}$/ regular expression"],
              "error": "Bad Request"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "التوكن مفقود/منتهي/غير صالح أو الـ role لا يطابق الـ endpoint",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "examples": {
              "missingToken": { "summary": "بدون توكن", "value": { "statusCode": 401, "message": "Unauthorized" } },
              "badRole": { "summary": "توكين لا يطابق الـ role", "value": { "statusCode": 401, "message": "جلسة غير صالحة" } }
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "integer", "example": 400 },
          "message": { "oneOf": [{ "type": "string" }, { "type": "array", "items": { "type": "string" } }] },
          "error": { "type": "string", "example": "Bad Request" }
        },
        "required": ["statusCode", "message"]
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "clv123abc" },
          "phone": { "type": "string", "example": "0591234567", "description": "رقم هاتف فلسطيني 10 أرقام يبدأ بـ 05" },
          "name": { "type": "string", "nullable": true, "example": "أحمد محمود" },
          "grade": { "type": "string", "nullable": true, "enum": ["GRADE_11", "GRADE_12"] },
          "branch": { "type": "string", "nullable": true, "enum": ["SCIENTIFIC", "LITERARY"] },
          "parentPhone": { "type": "string", "nullable": true, "example": "0599876543" },
          "role": { "type": "string", "enum": ["STUDENT", "TEACHER", "ADMIN"], "example": "STUDENT" },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "phone", "role", "createdAt"]
      },
      "Subscription": {
        "type": "object",
        "properties": {
          "planType": { "type": "string", "enum": ["TRIAL", "MONTHLY", "QUARTERLY", "YEARLY"] },
          "planName": { "type": "string", "example": "اشتراك شهري" },
          "endDate": { "type": "string", "format": "date-time" },
          "isActive": { "type": "boolean", "description": "true فقط إذا isActive && !isFrozen && endDate > now" },
          "videosPerSubject": { "type": "integer", "example": 10 }
        }
      },
      "MeResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/User" },
          {
            "type": "object",
            "properties": {
              "subscription": { "oneOf": [{ "$ref": "#/components/schemas/Subscription" }, { "type": "null" }] }
            }
          }
        ]
      },
...
      "StreamTokenResponse": {
        "type": "object",
        "properties": {
          "streamUrl": { "type": "string", "format": "uri", "nullable": true },
          "watermark": {
            "type": "object",
            "description": "اعرض name و phone على المشغّل لمنع التشارك",
            "properties": {
              "name": { "type": "string", "example": "أحمد محمود" },
              "phone": { "type": "string", "example": "0591234567" }
            }
          },
          "playbackRates": { "type": "array", "items": { "type": "number" }, "example": [0.75, 1, 1.25, 1.5, 2] },
          "qualities": { "type": "array", "items": { "type": "string" }, "example": ["auto", "360p", "480p", "720p", "1080p"] }
        }
      },
      "DownloadTokenResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string", "description": "توكين التنزيل (placeholder حالياً - يلزم ربطه بـ Signed URL من CDN لاحقاً)" },
          "expiresAt": { "type": "string", "format": "date-time" },
          "downloadDays": { "type": "integer", "example": 7 }
        }
      },
...
      "UpdateParentPhoneDto": {
        "type": "object",
        "required": ["parentPhone"],
        "properties": {
          "parentPhone": { "type": "string", "pattern": "^05\\d{8}$", "example": "0599876543" }
        }
      },
...
    }
  }
}